// backend/src/entities/enums/ErrorCode.ts
export enum ErrorCode {
	UNAUTHORIZED = "UNAUTHORIZED",
	INVALID_TOKEN = "INVALID_TOKEN",
	INTERNAL_ERROR = "INTERNAL_ERROR",
	PRODUCT_NOT_FOUND = "PRODUCT_NOT_FOUND",
	CATEGORY_NOT_FOUND = "CATEGORY_NOT_FOUND",
	INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
	VALIDATION_ERROR = "VALIDATION_ERROR",
	INVALID_INPUT = "INVALID_INPUT",
	PRODUCT_SEARCH_UNAVAILABLE = "PRODUCT_SEARCH_UNAVAILABLE",
	ADMIN_NOT_FOUND = "ADMIN_NOT_FOUND",
	ADDITIONAL_NOT_FOUND = "ADDITIONAL_NOT_FOUND",
	EDGE_NOT_FOUND = "EDGE_NOT_FOUND",
	SIZE_NOT_FOUND = "SIZE_NOT_FOUND",
	STORE_INFO_NOT_FOUND = "STORE_INFO_NOT_FOUND",
	INVALID_IMAGE = "INVALID_IMAGE",
	EMAIL_ALREADY_EXISTS = "EMAIL_ALREADY_EXISTS",
	USER_NOT_FOUND = "USER_NOT_FOUND",
	ORDER_NOT_FOUND = "ORDER_NOT_FOUND",
	PRODUCT_UNAVAILABLE = "PRODUCT_UNAVAILABLE",
	FAVORITE_NOT_FOUND = "FAVORITE_NOT_FOUND",
	FAVORITE_ALREADY_EXISTS = "FAVORITE_ALREADY_EXISTS",
	FORBIDDEN = "FORBIDDEN",
}